+++ test/tint/bug/tint/1725.wgsl.expected.ir.dxc.hlsl
@@ -10,7 +10,7 @@ void tint_symbol_1_inner(uint tint_symbol_2) {
   int tint_symbol_5 = int(0);
   uint v = 0u;
   tint_symbol.GetDimensions(v);
-  uint tint_symbol_6 = tint_symbol.Load((0u + (uint(min(tint_symbol_2, ((v / 4u) - 1u))) * 4u)));
+  uint tint_symbol_6 = tint_symbol.Load((0u + (min(tint_symbol_2, ((v / 4u) - 1u)) * 4u)));
 }
 
 [numthreads(1, 1, 1)]
