+++ test/tint/bug/tint/1875.wgsl.expected.ir.fxc.hlsl
@@ -2,7 +2,7 @@
 static uint count = 0u;
 RWByteAddressBuffer outputs : register(u1);
 void push_output(uint value) {
-  outputs.Store((0u + (uint(count) * 4u)), value);
+  outputs.Store((0u + (count * 4u)), value);
   count = (count + 1u);
 }
