+++ test/tint/ptr_ref/load/param/storage/vec2_f32_in_mat2x2.wgsl.expected.ir.dxc.hlsl
@@ -1,7 +1,7 @@
 
 ByteAddressBuffer S : register(t0);
 float2 func(uint pointer_indices[1]) {
-  return asfloat(S.Load2((0u + (uint(pointer_indices[0u]) * 8u))));
+  return asfloat(S.Load2((0u + (pointer_indices[0u] * 8u))));
 }
 
 [numthreads(1, 1, 1)]
