+++ test/tint/ptr_ref/load/param/storage/vec4_f32_in_mat2x4.wgsl.expected.ir.fxc.hlsl
@@ -1,7 +1,7 @@
 
 ByteAddressBuffer S : register(t0);
 float4 func(uint pointer_indices[1]) {
-  return asfloat(S.Load4((0u + (uint(pointer_indices[0u]) * 16u))));
+  return asfloat(S.Load4((0u + (pointer_indices[0u] * 16u))));
 }
 
 [numthreads(1, 1, 1)]
