+++ test/tint/ptr_ref/store/param/storage/struct_in_array.wgsl.expected.ir.fxc.hlsl
@@ -10,7 +10,7 @@ void v(uint offset, str obj) {
 
 void func(uint pointer_indices[1]) {
   str v_1 = (str)0;
-  v((0u + (uint(pointer_indices[0u]) * 4u)), v_1);
+  v((0u + (pointer_indices[0u] * 4u)), v_1);
 }
 
 [numthreads(1, 1, 1)]
