+++ test/tint/ptr_ref/store/param/storage/vec2_f32_in_mat2x2.wgsl.expected.ir.dxc.hlsl
@@ -1,7 +1,7 @@
 
 RWByteAddressBuffer S : register(u0);
 void func(uint pointer_indices[1]) {
-  S.Store2((0u + (uint(pointer_indices[0u]) * 8u)), asuint((0.0f).xx));
+  S.Store2((0u + (pointer_indices[0u] * 8u)), asuint((0.0f).xx));
 }
 
 [numthreads(1, 1, 1)]
