+++ test/tint/ptr_ref/store/param/storage/vec4_f32_in_mat2x4.wgsl.expected.ir.fxc.hlsl
@@ -1,7 +1,7 @@
 
 RWByteAddressBuffer S : register(u0);
 void func(uint pointer_indices[1]) {
-  S.Store4((0u + (uint(pointer_indices[0u]) * 16u)), asuint((0.0f).xxxx));
+  S.Store4((0u + (pointer_indices[0u] * 16u)), asuint((0.0f).xxxx));
 }
 
 [numthreads(1, 1, 1)]
