+++ test/tint/types/functions/shader_io/compute_subgroup_builtins_struct.wgsl.expected.ir.dxc.hlsl
@@ -6,7 +6,7 @@ struct ComputeInputs {
 
 RWByteAddressBuffer output : register(u0);
 void main_inner(ComputeInputs inputs) {
-  output.Store((0u + (uint(inputs.subgroup_invocation_id) * 4u)), inputs.subgroup_size);
+  output.Store((0u + (inputs.subgroup_invocation_id * 4u)), inputs.subgroup_size);
 }
 
 [numthreads(1, 1, 1)]
